+++ modules/moulinette-imagesearch.js
@@ -6,6 +6,7 @@ import { MoulinetteSearchResult } from "./moulinette-searchresult.js"
 export class MoulinetteImageSearch extends game.moulinette.applications.MoulinetteForgeModule {
 
   static SEARCH_BING_API = "https://api.bing.microsoft.com/v7.0/images/search"
+  static SEARCH_GOOGLE_API = "https://customsearch.googleapis.com/customsearch/v1"
   static SEARCH_CC_API = "https://api.openverse.engineering/v1/images"
   
   constructor() {
@@ -19,8 +20,9 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
    */
   async getPackList() {
     this.assetsPacks = []
+    this.assetsPacks.push({ idx: 3, special:"google", publisher: "Google", pubWebsite: "https://developers.google.com/custom-search/", name: "Google Custom Search v1", url: "https://developers.google.com/custom-search/v1/overview", license: "depends", isRemote: true })
     this.assetsPacks.push({ idx: 1, special:"bing", publisher: "Microsoft", pubWebsite: "https://microsoft.com", name: "Bing Search v7.0", url: "http://bing.com/", license: "depends", isRemote: true })
-    this.assetsPacks.push({ idx: 2, special:"cc", publisher: "Creative Commons", pubWebsite: "https://opensource.creativecommons.org/", name: "CC Search v1.0", url: "https://opensource.creativecommons.org/archives/cc-search/", license: "depends", isRemote: true })    
+    this.assetsPacks.push({ idx: 2, special:"cc", publisher: "Creative Commons", pubWebsite: "https://opensource.creativecommons.org/", name: "CC Search v1.0", url: "https://opensource.creativecommons.org/archives/cc-search/", license: "depends", isRemote: true })
     return duplicate(this.assetsPacks)
   }
   
@@ -50,7 +52,60 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
     let data = await response.json()
     
     let results = []
-    data.value.forEach( r => results.push({ src: "Microsoft Bing", name: r.name, thumb: r.thumbnailUrl, url: r.contentUrl, page: r.hostPageUrl, width: r.width, height: r.height, format: r.encodingFormat}));
+    data.value.forEach( r => results.push({
+      src: "Microsoft Bing",
+      name: r.name,
+      thumb: r.thumbnailUrl,
+      url: r.contentUrl,
+      page: r.hostPageUrl,
+      width: r.width,
+      height: r.height,
+      format: r.encodingFormat}));
+    return results
+  }
+
+  /**
+  /* Google Search implementation
+   */
+  async searchGoogle(searchTerms, googleKey, googleCx) {
+    // execute search
+    const params = new URLSearchParams({
+      key: googleKey,
+      cx: googleCx,
+      q: searchTerms,
+      searchType: "image",
+      num: 10,
+      start: 1,
+      filter: 1
+    })
+
+    let results = []
+
+    // retrieve 30 results (= 5 requests)
+    for(let idx=0; idx<3; idx++) {
+
+      params.set("start", idx * 10 + 1)
+      const response = await fetch(`${MoulinetteImageSearch.SEARCH_GOOGLE_API}?${params}`).catch(function(e) {
+        console.log(`MoulinetteClient | Cannot establish connection to server ${MoulinetteImageSearch.SEARCH_GOOGLE_API}`, e)
+      });
+
+      if( !response || response.status != 200 ) {
+        console.error("MoulinetteImageSearch | Invalid response from Google API", response)
+        return [];
+      }
+      let data = await response.json()
+
+      data.items.forEach( r => results.push({
+        src: "Google Search",
+        name: r.title,
+        thumb: r.image.thumbnailLink,
+        url: r.link,
+        page: r.image.contextLink,
+        width: r.image.width,
+        height: r.image.height,
+        format: r.fileFormat.split("/").pop()}));
+    }
+
     return results
   }
   
@@ -64,7 +119,7 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
     }
     const params = new URLSearchParams({
       q: searchTerms,
-      page_size: 150
+      page_size: 20
     })
     
     const response = await fetch(`${MoulinetteImageSearch.SEARCH_CC_API}/?format=json&${params}`, header).catch(function(e) {
@@ -80,7 +135,16 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
     let results = []
     data.results.forEach( r => {
       const format = r.url.substring(r.url.lastIndexOf(".")+1)
-      results.push({ src: "Creative Commons", name: r.title, thumb: r.thumbnail, url: r.url, license: r.license, licenseUrl: r.license_url, page: r.foreign_landing_url, noSize: true, format: format })
+      results.push({
+        src: "Creative Commons",
+        name: r.title,
+        thumb: r.thumbnail,
+        url: r.url,
+        license: r.license,
+        licenseUrl: r.license_url,
+        page: r.foreign_landing_url,
+        noSize: true,
+        format: format })
     });
     return results
   }
@@ -91,10 +155,26 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
   async getAssetList(searchTerms, packId) {
     let assets = []
  
+    // error handling
     let pack = packId ? this.assetsPacks.find(p => p.idx == packId) : null
     const bingKey = game.settings.get("moulinette-imagesearch", "bing-key")
     if(pack && pack.special == "bing" && (!bingKey | bingKey.length == 0)) {
       assets.push(`<div class="error">${game.i18n.localize("mtte.noBingKey")}</div>`)
+      return assets;
+    }
+    const googleKey = game.settings.get("moulinette-imagesearch", "google-key")
+    if(pack && pack.special == "google" && (!googleKey | googleKey.length == 0)) {
+      assets.push(`<div class="error">${game.i18n.localize("mtte.noGoogleKey")}</div>`)
+      return assets;
+    }
+    const googleEngineId = game.settings.get("moulinette-imagesearch", "google-engine-id")
+    if(pack && pack.special == "google" && (!googleEngineId | googleEngineId.length == 0)) {
+      assets.push(`<div class="error">${game.i18n.localize("mtte.noGoogleEngine")}</div>`)
+      return assets;
+    }
+    const openverseEnabled = game.settings.get("moulinette-imagesearch", "openverse-enabled")
+    if(pack && pack.special == "cc" && !openverseEnabled) {
+      assets.push(`<div class="error">${game.i18n.localize("mtte.openverseDisabled")}</div>`)
       return assets;
     }
     
@@ -108,7 +188,10 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
     if((!pack || pack.special == "bing") && bingKey && bingKey.length > 0) {
       this.searchResults.push(...await this.searchBing(searchTerms, bingKey))
     }
-    if(!pack || pack.special == "cc") {
+    if((!pack || pack.special == "google") && googleKey && googleKey.length > 0 && googleEngineId && googleEngineId.length > 0) {
+      this.searchResults.push(...await this.searchGoogle(searchTerms, googleKey, googleEngineId))
+    }
+    if((!pack || pack.special == "cc") && openverseEnabled) {
       this.searchResults.push(...await this.searchCC(searchTerms))
     }
     
@@ -120,6 +203,8 @@ export class MoulinetteImageSearch extends game.moulinette.applications.Moulinet
       idx++
       assets.push(`<div class="imageresult draggable" title="${r.name}" data-idx="${idx}"><img width="100" height="100" src="${r.thumb}"/></div>`)
     })
+
+    assets.push(`<div class="text">Hello</div>`)
   
     return assets
   }