+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.1.0] - 2022-09-24
+### Fixed
+- Openverse max number of results is now 20
+### Added
+- Google Search
+
 ## [10.0.2] - 2022-09-04
 ### Fixed
 - 10.0.1: module packaging fix